Version 1.2
Added nav bar (not functional, but it does highlight the current page and whatever the user is hovering over) and cleaned up the site even more. Further work on the nav bar and other pages will commence once the welcome page is fully fleshed out, but so far it's looking good. Also added another line to the welcome page text - currently waiting for stakeholder to provide more info on what exactly they want in the welcome page, and maybe as well provide examples on some images they may want to be on the site too